+++ library/SolverUtils/DriverParareal.h
@@ -102,27 +102,6 @@ private:
 
     void WriteTimeChunkOuput(void);
 
-    void SpeedUpAnalysis();
-
-    void PrintSpeedUp(NekDouble fineSolveTime, NekDouble coarseSolveTime,
-                      NekDouble restTime, NekDouble interTime,
-                      NekDouble commTime, NekDouble predictorTime);
-
-    NekDouble ComputeSpeedUp(const size_t iter, NekDouble fineSolveTime,
-                             NekDouble coarseSolveTime, NekDouble restTime,
-                             NekDouble interTime, NekDouble commTime,
-                             NekDouble predictorTime);
-
-    NekDouble EstimateCommunicationTime(void);
-
-    NekDouble EstimateRestrictionTime(void);
-
-    NekDouble EstimateInterpolationTime(void);
-
-    NekDouble EstimateSolverTime(const size_t timeLevel, const size_t nstep);
-
-    NekDouble EstimatePredictorTime(void);
-
     static constexpr size_t m_fineLevel   = 0;
     static constexpr size_t m_coarseLevel = 1;
     Array<OneD, Array<OneD, NekDouble>> m_initialCondition;